Proposal for different packet versions handling.
This proposal provides to korangar different version of packet handling,
leveraging existing handler mecanism.
For packet emission, a new "factory" concept has been added.

Goal achieved by this proposal:
- Korangar is unaware of packet version
- There at the moment no condition on packet version
- Korangar is now extensible by any 3rd party developer who want to use
  its own packet version. Support of a specific version of packet can be
  done in an external crate, one can maintain its own crate of packet
  handling and inject its implementation in korangar main without any
  changes.

Packet Factory
The concept of factory is to map game data into packet ready to be sent
to the wire.
Factory are traits to be implemented to emit packet